feat(ci): add mypy, ruff format, calibration-by-default, and calibrate command
Four gaps addressed:

CI (python.yml):
- Add ruff format --check step so formatting violations fail the build
- Add mypy strict type-check step to enforce the existing pyproject config

Type correctness (mypy strict, 0 errors):
- Add FeatureExtractor import to pipeline and type extractor param properly
- Remove stale type: ignore[union-attr] comments made redundant by proper typing
- Add model_id @property to Model Protocol so CalibratedModel satisfies it
- Fix _compute_clv closing variable declared as float | None to satisfy narrowing
- Fix _weather_kwargs return type to dict[str, object]
- Type _EXTRACTOR_FACTORIES and _build_model_and_extractor return in cli.py
- Add sklearn/scipy mypy overrides (no stubs available for these packages)
- Wrap elo win-probability return in float() to satisfy no-any-return

Calibration by default (architecture compliance):
- Add _maybe_calibrate(model, calibrate=True) helper in cli.py
- Add --calibrate/--no-calibrate flag to bet backtest (default: on)
- Architecture rule "calibration always applied before value analysis" now
  enforced by default rather than being opt-in via CalibratedModel wrapping

New CLI command (bet calibrate):
- Runs walk-forward prediction loop with min_edge=0.0 to collect all predictions
- Reports Brier score, log-loss, and ECE for independent calibration evaluation
- Accepts same --sport/--data/--model/--min-train flags as bet backtest

Tests: 337 passing (9 new in tests/test_cli_calibration.py)

+++ src/bet/modeling/elo.py
@@ -65,7 +65,7 @@ class EloModel:
         Returns:
             Probability in (0, 1) that team A wins.
         """
-        return 1.0 / (1.0 + 10.0 ** ((rating_b - rating_a) / 400.0))
+        return float(1.0 / (1.0 + 10.0 ** ((rating_b - rating_a) / 400.0)))
 
     def _mov_multiplier(self, home_score: int, away_score: int) -> float:
         """Scale K by log margin so blowouts carry more information than squeakers.
